feat(pool): add new param wait_for_connection
Supports:
- ConnectionPool(wait_for_connection: int = 0)
- ClusterConnectionPool(wait_for_connection: int = 0)
and anything which initializes a pool:
- StrictRedis(wait_for_connection: int = 0
- StrictRedisCluster(wait_for_connection: int = 0)
- SentinelConnectionPool(wait_for_connection: int = 0)
- etc...

Which fixes #51 by allowing an optional wait_for_connection parameter
which can be used to control the maximum amount of time the connection
pool will wait for a connection before exiting.

This allows the client to decide the behaviour of the connection pool,
as you described in the original issue. The default behaviour has not
changed.