fix bug: do not use stream_timeout in the PubSubWorkerThread
If we don't give `get_message()` a timeout and the connection was
created with a `stream_timeout`, then the `get_message()` will timeout
and the thread stops waiting for more messages.

Use a default timeout of 1.0, which can be changed with a constructor
parameter, to fix that and also allows the thread to exit cleanly if
no timeouts are given, since it won't block forever.

Fixes #106