Add an actual simple example!
The 'simple' example seemed to have every possible feature crammed in!

Aside from my conviction that the word 'simple' should really never be used in documentation (if something is simple then why are you documenting it?), my main difficulty here was trying to figure out where `async` and `await` are needed on the various statements.
i.e. if I'm just queueing up commands for execution, why do I need to await them?  And why do I need to await the call to `client.pipeline()`?  This is all quite verbose and certainly not simple.
Also I emphatically assummed that the `with` context manager would actually execute the pipeline upon exit, otherwise what is a context manager for??  (yes I've read the source and see that you can reuse a pipeline; I would have just created a new pipeline for a new transaction)

Anyhow hopefully this smaller example is correct and will be helpful to future readers of the documentation.